+++ src/App.tsx
@@ -1,32 +1,10 @@
 import './App.scss';
-import { NavLink, Outlet } from 'react-router-dom';
-import cn from 'classnames';
-
-const getActiveClass = ({ isActive }: { isActive: boolean }) =>
-  cn('navbar-item', {
-    'has-background-grey-lighter': isActive,
-  });
+import { Outlet } from 'react-router-dom';
+import { Navigation } from './components/Navigation';
 
 export const App = () => (
   <div data-cy="app">
-    <nav
-      data-cy="nav"
-      className="navbar is-fixed-top has-shadow"
-      role="navigation"
-      aria-label="main navigation"
-    >
-      <div className="container">
-        <div className="navbar-brand">
-          <NavLink className={getActiveClass} to="/">
-            Home
-          </NavLink>
-
-          <NavLink className={getActiveClass} to="/people">
-            People
-          </NavLink>
-        </div>
-      </div>
-    </nav>
+    <Navigation />
 
     <main className="section">
       <Outlet />
